# services/api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.OrderItemRequest'
        type: array
    type: object
  handler.CreatePartRequest:
    properties:
      brand:
        type: string
      category_id:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.OrderItemRequest:
    properties:
      part_id:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  handler.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  handler.UpdateStatusRequest:
    properties:
      status:
        type: integer
    type: object
  handler.UpdateStockRequest:
    properties:
      part_id:
        type: string
      quantity:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway for AutoParts Hub
  title: AutoParts Hub API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /catalog/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - catalog
  /catalog/parts:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: Brand filter
        in: query
        name: brand
        type: string
      - description: Car make
        in: query
        name: make
        type: string
      - description: Car model
        in: query
        name: model
        type: string
      - description: Car year
        in: query
        name: year
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search parts
      tags:
      - catalog
    post:
      consumes:
      - application/json
      parameters:
      - description: Create part request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create part
      tags:
      - catalog
  /catalog/parts/{id}:
    get:
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get part by ID
      tags:
      - catalog
  /inventory/stock:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update stock
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update stock quantity
      tags:
      - inventory
  /inventory/stock/{part_id}:
    get:
      parameters:
      - description: Part ID
        in: path
        name: part_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get stock by part ID
      tags:
      - inventory
  /orders:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Create order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - orders
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"